refactor(mihomo): rename connection_manager to ws_connection_manager
