+++ crates/tauri-plugin-mihomo/src/ws_connection_manager.rs
